+++ spec/cases/graph_api_batch_spec.rb
@@ -516,36 +516,34 @@ describe "Koala::Facebook::GraphAPI in batch mode" do
         batch_api.get_connections(@app_id, 'insights', {}, {"access_token" => @app_api.access_token})
       end
       me['id'].should_not be_nil
-      insights.should be_an(Array)
+      insights.should be_an(Koala::Facebook::GraphCollection)
     end
 
-    it "handles batch operation-specific token when paging" do
-
+    it "preserves batch-op specific access tokens in GraphCollection returned from batch" do
       # Provide an alternate token for a batch operation
-      @args = { 'access_token' => @app_api.access_token }
+      @other_access_token_args = { 'access_token' => @app_api.access_token }
 
+      # make a batch call for insights using another token
       me, insights = @api.batch do |batch_api|
         batch_api.get_object('me')
-        batch_api.get_connections(@app_id, 'insights', {}, @args)
+        batch_api.get_connections(@app_id, 'insights', {}, @other_access_token_args)
       end
-
-      insights.should be_a(Koala::Facebook::GraphCollection)
 
       # The alternate token is returned with the next page parameters
       # The GraphCollection should receive a request for the next_page_params during paging
-      insights.should_receive(:next_page_params).and_return([stub("base"), @args.dup])
+      insights.should_receive(:next_page_params).and_return([stub("base"), @other_access_token_args.dup])
 
+      # The alternate access token should pass through to making the request
       # Koala should receive a request during paging using the alternate token
       Koala.should_receive(:make_request).with(
           anything,
-          hash_including(@args.dup),
+          hash_including(@other_access_token_args.dup),
           anything,
           anything
       ).and_return(Koala::HTTPService::Response.new(200, "", ""))
 
       # Page the collection
       insights.next_page
-
     end
 
     it "inserts errors in the appropriate place, without breaking other results" do